\encoding{UTF-8}

\name{InVVRefdata}

\alias{InVVRefdata}

\title{Input/Edit In Vivo raw data for IV, oral solution or IR drugs}

\description{
   Users input/edit In Vivo raw data for IV, oral solution or IR drugs.
   Loading data file which was saved as .csv form or .RData form, or key in data in the data editor window.

}
\keyword{misc}