\name{InVVTestdata}

\alias{InVVTestdata}

\title{Input/Edit In Vivo raw data for extended release formulations with different rates and Edit In Vitro Dissoution Data}

\description{
   Users input/edit edit In Vivo raw data for extended release formulations with different rates and In Vitro Dissoution Data.
   Loading data file which was saved as .csv form or .RData form, or key in data in the data editor window.

}
\keyword{misc}