\encoding{UTF-8}

\name{ivivc_all_plots}

\alias{ivivc_all_plots}

\title{to generate all plots for ivivc runs}

\description{to generate all plots for ivivc runs

Authors: Hsin-ya Lee, Yung-jin Lee
Kaohsiung Veterans General Hospital (HY) &
ptpc inc. (YJ),
Kaoshiung, Taiwan 80708
E-mail: mobilepk@gmail.com
website: http://pkpd.kmu.edu.tw/ivivc
}

\keyword{misc}