\encoding{UTF-8}

\name{ivivc_demo}

\alias{ivivc_demo}

\title{ivivc Demo}

\description{
   1.iv data:
     Dose=200 mg
     Vd=35 L
     Ke=0.0863 h-1
   2.in vitro dissolution (FRD) and in vivo observed plasma conc.(conc.obs)
   Time=c(0,1,2,3,4,5,6,7,8)
   FRD=c(0,18.7,45.3,62.4,77.3,84.9,91.5,92.7,95.8)
   conc.obs=c(0,0.891,1.997,2.616,3.411,3.33,3.868,3.371,3.433)

}

\references{
 C. J. Shishoo, S. S. Savale, S. A. Shah, , I. S. Rathod,  and P. K. Mukherjee,
 Indian J.Pharm. Sci., 2002, 64, 222.
}

\keyword{misc}