\encoding{UTF-8}

\name{logo_plot_desc}

\alias{logo_plot_desc}

\title{Description for ivivc for R plots version}

\description{Description for ivivc for R plots version

Authors: Hsin-ya Lee, Yung-jin Lee
Kaohsiung Veterans General Hospital (HY) &
ptpc inc. (YJ),
Kaoshiung, Taiwan 80794
E-mail: mobilepk@gmail.com
website: http://pkpd.kmu.edu.tw/bear
}

\keyword{misc}