\encoding{UTF-8}

\name{noniv.route}

\alias{noniv.route}

\title{Options for Extravascular Models Menu}

\description{
  Provide the following three PK models for selection:
   1. one-compartment, single-dose, and first-ordered absorption with lag time model;
   2. one-compartment, single-dose, and first-ordered absorption without lag time model;
   3. one-compartment, single-dose, and zero-ordered absorption without lag time model;
  
}

\keyword{misc}

