\encoding{UTF-8}

\name{WagNel}

\alias{WagNel}

\title{Wagner-Nelson method}

\description{
   Nine functions in Wagner-Nelson method:
   1.calculate AUC(0~t)
   2.calculate Fab (%)(cumulative absorption fraction)
   3.perform a Linear Regression
   4.calculate FABpred(%)(cumulative predicted absorption fraction)
   5.calculate conc.pred (predicted concentration)
   6.PECmax(Average absolute prediction error of Cmax)
   7.PEAUC (Average absolute prediction error of AUC)
   8.Plot In vivo Absorption : Fraction of Absorption(%) vs. time
   9.In vivo plasma concentration (Predicted): Plasma conc. vs. time
}

\references{
  Dr.Mukesh Gohel * , R. R. Delvadia, D. C. Parikh, M. M. Zinzuwadia, C. D. Soni, K. G. Sarvaiya
  R. Joshi and A. S. Dabhi.  Simplified Mathematical Approach for Back Calculation in Wagner-Nelson Method. 
  http://www.pharmainfo.net/reviews/simplified-mathematical-approach-back-calculation-wagner-nelson-method 
} 

\keyword{misc}